Add GitHub Actions to run tests (#121)
There are two separate actions - one for the mandatory tests, and one
for the extra tests which is only run if any extra tests were modified.

There's also a fork of someone else's GitHub Jest reporter in here to
post in-line comments with test failures. Unfortunately, this has been
pending merge into Jest for quite a while, so if we want to use it we
need to fork it.

It could be nice, rather than forking it into each of our weeks of
coursework, to release the code as a package and depend on it, but I'm
not sure about the license story of doing so.